+++ src/FreeElephants/AltEra/Exception/AltEraExceptionInterface.php
@@ -3,10 +3,8 @@
 namespace FreeElephants\AltEra\Exception;
 
 /**
- * TODO rename to AltEraEception.
- *
  * @author samizdam
  */
-interface CalendarExceptionInterface
+interface AltEraExceptionInterface
 {
 }