+++ src/FreeElephants/AltEra/Exception/DomainException.php
@@ -5,6 +5,6 @@ namespace FreeElephants\AltEra\Exception;
 /**
  * @author samizdam
  */
-class DomainException extends \DomainException implements CalendarExceptionInterface
+class DomainException extends \DomainException implements AltEraExceptionInterface
 {
 }