+++ src/FreeElephants/AltEra/Exception/LogicException.php
@@ -5,6 +5,6 @@ namespace FreeElephants\AltEra\Exception;
 /**
  * @author samizdam
  */
-class LogicException extends \LogicException implements CalendarExceptionInterface
+class LogicException extends \LogicException implements AltEraExceptionInterface
 {
 }